backend/requirements.txt
# ==========================================
# FastAPI 核心依赖
# ==========================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0

# ==========================================
# 数据库相关
# ==========================================
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.11

# ==========================================
# 数据验证和配置
# ==========================================
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
python-dotenv==1.2.1

# ==========================================
# 认证和安全
# ==========================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==5.0.0
cryptography==46.0.3

# ==========================================
# 日志记录
# ==========================================
loguru==0.7.2

# ==========================================
# HTTP 客户端
# ==========================================
httpx==0.25.2
requests==2.32.5
aiohttp==3.13.3

# ==========================================
# 工具库
# ==========================================
python-dateutil==2.8.2

# ==========================================
# 可选依赖
# ==========================================
redis==5.0.1

# ==========================================
# 开发工具（可选，仅用于开发环境）
# ==========================================
# pytest==7.4.3
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0
# black==23.11.0
# flake8==6.1.0
# isort==5.12.0
# mypy==1.7.1
# pre-commit==3.5.0